fixes issue with header->sig not being the right size.   This kept things from executing (how did this ever work?) updates openssl updates boost fixes return type of astNode::infixPriority && stackPriority() to be size_t fixes debugAdapter setFunctionBreakpoints && setBreakpoints to check if LastInstance is valid

+++ compilerAST/astNodeParserSupport.cpp
@@ -6,7 +6,7 @@
 
 #include "compilerParser/fileParser.h"
 
-int astNode::infixPriority()
+size_t astNode::infixPriority()
 {
 	if ( opToOpDef[(size_t)op] )
 	{
@@ -14,7 +14,7 @@ int astNode::infixPriority()
 	} else return 1;
 }
 
-int astNode::stackPriority()
+size_t astNode::stackPriority()
 {
 	if ( opToOpDef[(size_t)op] )
 	{
